# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cy-performance API",
    "description": "Cycling analysis + training plan backend (read models, what-ifs, jobs).",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/fitness": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Fitness",
        "description": "Daily load and CTL/ATL/TSB (default: last 120 days).",
        "operationId": "get_fitness_v1_fitness_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitnessSeries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/trends/recompute": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Recompute Trends",
        "description": "Recompute and store the trends report (~100 ms).",
        "operationId": "recompute_trends_v1_trends_recompute_post",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Recompute Trends V1 Trends Recompute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/explain/{key}": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Explanation",
        "description": "The persisted Explanation for ``key`` (e.g. ``ride:42``, ``readiness.2026-10-06``).",
        "operationId": "get_explanation_v1_explain__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/glossary/{term}": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Glossary",
        "description": "One glossary page as Markdown (zh-TW).",
        "operationId": "get_glossary_v1_glossary__term__get",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Term"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/readiness": {
      "get": {
        "tags": [
          "readiness"
        ],
        "summary": "History",
        "description": "Stored verdicts (default: last 28 days).",
        "operationId": "history_v1_readiness_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadinessOut"
                  },
                  "title": "Response History V1 Readiness Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/readiness/{day}": {
      "get": {
        "tags": [
          "readiness"
        ],
        "summary": "Get Day",
        "description": "Stored verdict, or computed on the fly when none is stored.",
        "operationId": "get_day_v1_readiness__day__get",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/readiness/recompute": {
      "post": {
        "tags": [
          "readiness"
        ],
        "summary": "Recompute",
        "description": "Score and store the given days (default: today).",
        "operationId": "recompute_v1_readiness_recompute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_recompute_v1_readiness_recompute_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReadinessOut"
                  },
                  "type": "array",
                  "title": "Response Recompute V1 Readiness Recompute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plan": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "Stored",
        "description": "Stored proposals (what the calendar will get), rendered with steps and workout text.",
        "operationId": "stored_v1_plan_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 56,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlannedDayOut"
                  },
                  "title": "Response Stored V1 Plan Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plan/preview": {
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Preview",
        "description": "What-if plan (availability, days off, indoor days, readiness, bias, CTL); nothing stored.",
        "operationId": "preview_v1_plan_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PlanPreviewRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plan/commit": {
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Commit",
        "description": "Replan and store proposals (never writes to intervals.icu); 409 while a run is active.",
        "operationId": "commit_v1_plan_commit_post",
        "parameters": [
          {
            "name": "today",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Today"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_recompute_v1_readiness_recompute_post": {
        "properties": {
          "dates": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "date"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dates"
          }
        },
        "type": "object",
        "title": "Body_recompute_v1_readiness_recompute_post"
      },
...
      "ExplainOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "explanation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "source",
          "explanation"
        ],
        "title": "ExplainOut",
        "description": "``GET /v1/explain/{key}``."
      },
...
      "FitnessPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "load": {
            "type": "number",
            "title": "Load",
            "description": "Actual daily load (all sports)"
          },
          "ctl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctl"
          },
          "atl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atl"
          },
          "tsb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tsb"
          },
          "source": {
            "type": "string",
            "enum": [
              "icu",
              "sim",
              "none"
            ],
            "title": "Source",
            "default": "none"
          },
          "ctl_sim": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctl Sim"
          },
          "acwr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acwr"
          },
          "monotony": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monotony"
          },
          "planned_load": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Load",
            "description": "Our proposed load for that day"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "load"
        ],
        "title": "FitnessPoint",
        "description": "One day of the performance-management chart."
      },
      "FitnessSeries": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/FitnessPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start",
          "end",
          "points"
        ],
        "title": "FitnessSeries",
        "description": "``GET /v1/fitness``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanOut": {
        "properties": {
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          },
          "season_key": {
            "type": "string",
            "title": "Season Key"
          },
          "persisted": {
            "type": "boolean",
            "title": "Persisted"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/PlannedDayOut"
            },
            "type": "array",
            "title": "Days"
          },
          "weeks": {
            "items": {
              "$ref": "#/components/schemas/WeekTargetOut"
            },
            "type": "array",
            "title": "Weeks"
          },
          "adaptations": {
            "items": {
              "$ref": "#/components/schemas/Explanation"
            },
            "type": "array",
            "title": "Adaptations"
          },
          "repairs": {
            "items": {
              "$ref": "#/components/schemas/Explanation"
            },
            "type": "array",
            "title": "Repairs"
          },
          "violations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Violations"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review"
          },
          "projection": {
            "items": {
              "$ref": "#/components/schemas/FitnessPoint"
            },
            "type": "array",
            "title": "Projection",
            "description": "Simulated CTL/ATL/TSB over the horizon"
          },
          "changes": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Changes"
          },
          "compute_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compute Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "today",
          "season_key",
          "persisted",
          "days",
          "weeks",
          "adaptations",
          "repairs",
          "violations",
          "needs_review"
        ],
        "title": "PlanOut",
        "description": "``GET /v1/plan`` and ``POST /v1/plan/preview``."
      },
      "PlanPreviewRequest": {
        "properties": {
          "today": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Today"
          },
          "weekday_minutes": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "mon",
                "tue",
                "wed",
                "thu",
                "fri",
                "sat",
                "sun"
              ]
            },
            "type": "object",
            "title": "Weekday Minutes"
          },
          "weekly_max_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2400.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Max Minutes"
          },
          "date_minutes": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "format": "date"
            },
            "type": "object",
            "title": "Date Minutes",
            "description": "Cap minutes on dates; 0 = day off"
          },
          "indoor_days": {
            "items": {
              "type": "string",
              "format": "date"
            },
            "type": "array",
            "title": "Indoor Days"
          },
          "readiness": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "REST",
                  "EASY",
                  "AS_PLANNED",
                  "UPGRADE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness"
          },
          "bias": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bias"
          },
          "horizon_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 56.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon Days"
          },
          "ctl": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 200.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctl"
          },
          "atl": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 300.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Atl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PlanPreviewRequest",
        "description": "What-if inputs for ``POST /v1/plan/preview`` (nothing is stored)."
      },
...
      "PlannedDayOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "role_zh": {
            "type": "string",
            "title": "Role Zh"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "name_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Zh"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "tss": {
            "type": "number",
            "title": "Tss"
          },
          "minutes": {
            "type": "integer",
            "title": "Minutes"
          },
          "max_minutes": {
            "type": "integer",
            "title": "Max Minutes"
          },
          "outdoor": {
            "type": "boolean",
            "title": "Outdoor"
          },
          "params": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Params"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PlanStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "workout_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Text"
          },
          "note_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Zh"
          },
          "mutable": {
            "type": "boolean",
            "title": "Mutable",
            "default": true
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Explanation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "role",
          "role_zh",
          "intent",
          "tss",
          "minutes",
          "max_minutes",
          "outdoor"
        ],
        "title": "PlannedDayOut",
        "description": "One day of the horizon."
      },
...
      "ReadinessOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "recommendation_zh": {
            "type": "string",
            "title": "Recommendation Zh"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ReadinessComponentOut"
            },
            "type": "array",
            "title": "Components"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          },
          "rule_hits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rule Hits"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Explanation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "score",
          "status",
          "recommendation",
          "recommendation_zh",
          "components",
          "missing",
          "rule_hits"
        ],
        "title": "ReadinessOut",
        "description": "``GET /v1/readiness/{date}``."
      },
...
    }
  }
}